Refactor/abstract app into models, routes, schemas, seeds files
