+++ .gitignore
@@ -1,3 +1,6 @@
 /test/__pycache__
 app/__pycache__
 env
+coverage
+.pytest_cache
+test/__pycache__
