+++ app/models.py
@@ -1,0 +1,79 @@
+from app import *
+
+class ConnectionStatus(enum.Enum):
+    APPROVED = "Approved"
+    PENDING = "Pending"
+    REJECTED = "Rejected"
+
+class PlaysOrNeeds(enum.Enum):
+    PLAYS = "Plays"
+    NEEDS = "Needs"
+
+user_instrument = db.Table('user_instrument',
+    db.Column('user_id', db.Integer, db.ForeignKey('user.id'), primary_key=True),
+    db.Column('instrument_id', db.Integer, db.ForeignKey('instrument.id'), primary_key=True)
+)
+
+user_needs_instrument = db.Table('user_needs_instrument',
+    db.Column('user_id', db.Integer, db.ForeignKey('user.id'), primary_key=True),
+    db.Column('needs_instrument_id', db.Integer, db.ForeignKey('needs_instrument.id'), primary_key=True)
+)
+
+user_genre = db.Table('user_genre',
+    db.Column('user_id', db.Integer, db.ForeignKey('user.id'), primary_key=True),
+    db.Column('genre_id', db.Integer, db.ForeignKey('genre.id'), primary_key=True)
+)
+
+user_connection = db.Table('user_connection',
+    db.Column('user_id', db.Integer, db.ForeignKey('user.id'), primary_key=True),
+    db.Column('friend_id', db.Integer, db.ForeignKey('user.id'), primary_key=True),
+    db.Column('status', db.Enum(ConnectionStatus))
+)
+
+class User(db.Model):
+    __tablename__ = "user"
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(100))
+    display_email = db.Column(db.String(255))
+    picture_url = db.Column(db.String(255))
+    about = db.Column(db.String(255))
+    zipcode = db.Column(db.String(5))
+    date_created = db.Column(db.DateTime, default=datetime.utcnow)
+    instruments = db.relationship('Instrument', secondary=user_instrument,  lazy='select', backref=db.backref('users', lazy=True))
+    needs_instruments = db.relationship('NeedsInstrument', secondary=user_needs_instrument,  lazy='joined', backref=db.backref('users', lazy=True))
+    genres = db.relationship('Genre', secondary=user_genre, lazy='dynamic', backref=db.backref('users', lazy=True))
+    connections = db.relationship('User', secondary=user_connection, primaryjoin=id == user_connection.c.user_id, secondaryjoin=id == user_connection.c.friend_id, lazy='dynamic', backref=db.backref('users', lazy=True))
+
+    def __init__(self, name, display_email, picture_url, about, zipcode):
+        self.name = name
+        self.display_email = display_email
+        self.picture_url = picture_url
+        self.about = about
+        self.zipcode = zipcode
+
+class Instrument(db.Model):
+    __tablename__ = "instrument"
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String)
+    date_created = db.Column(db.DateTime, default=datetime.utcnow)
+
+    def __init__(self, name):
+        self.name = name
+
+class NeedsInstrument(db.Model):
+    __tablename__ = "needs_instrument"
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String)
+    date_created = db.Column(db.DateTime, default=datetime.utcnow)
+
+    def __init__(self, name):
+        self.name = name
+
+class Genre(db.Model):
+    __tablename__ = "genre"
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String)
+    date_created = db.Column(db.DateTime, default=datetime.utcnow)
+
+    def __init__(self, name):
+        self.name = name
