+++ requirements.txt
@@ -8,3 +8,4 @@ gunicorn
 marshmallow-jsonapi
 flask-cors
 pgeocode
+coverage
